Fix warnings by safely downcasting Py_ssize_t to int

+++ bson/_cbsonmodule.h
@@ -75,7 +75,7 @@ typedef struct codec_options_t {
 /* C API functions */
 #define _cbson_buffer_write_bytes_INDEX 0
 #define _cbson_buffer_write_bytes_RETURN int
-#define _cbson_buffer_write_bytes_PROTO (buffer_t buffer, const char* data, int size)
+#define _cbson_buffer_write_bytes_PROTO (buffer_t buffer, const char* data, Py_ssize_t size)
 
 #define _cbson_write_dict_INDEX 1
 #define _cbson_write_dict_RETURN int
